first final build
need more tests
need api descr

+++ src/main/java/com/degenerates/memium/controller/LikeOpenController.java
@@ -16,7 +16,7 @@ public class LikeOpenController {
     LikeFacade likeFacade;
 
     @GetMapping("/{articleId}")
-    public ResponseEntity<Integer> getLikeForArticle(@PathVariable UUID articleId) {
+    public ResponseEntity<Long> getLikeForArticle(@PathVariable UUID articleId) {
         return likeFacade.getLikeAmountForArticle(articleId);
     }
 }
